connectaflow/backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
sqlmodel
pydantic-settings
google-generativeai
pandas
openpyxl
requests
asyncpg
alembic
duckduckgo-search
polars
loguru
orjson
litellm
instructor
pypdf
python-docx
python-pptx
reportlab
